+++ migrations/3-down.sql
@@ -1,0 +1,3 @@
+UPDATE `${wpdb_prefix}postmeta`
+SET `meta_key` = "eddbk_session_lengths"
+WHERE `meta_key` = "eddbk_session_types";