+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [[*next-version*]] - YYYY-MM-DD
+### Added
+- New migration for DB version 3.
+- More migration events for more granular migration detection and handling.
 
 ## [0.1-alpha11] - 2018-08-28
 ### Changed